Adapt features contexts and move fixtures to seeds in new AcmeAppBundle
