+++ .travis.yml
@@ -12,7 +12,7 @@ before_script:
   - php Tests/Functionnal/bin/console --env=test cache:warmup
   - php Tests/Functionnal/bin/console --env=test doctrine:schema:create
   - php Tests/Functionnal/bin/console --env=test victoire:generate:view
-  - php Tests/Functionnal/bin/console --env=test doctrine:fixtures:load --fixtures="Bundle/CoreBundle/DataFixtures/ORM" -n --append
+  - php Tests/Functionnal/bin/console --env=test doctrine:fixtures:load --fixtures="Tests/Functionnal/src/Acme/AppBundle/DataFixtures/ORM" -n --append
   - php Tests/Functionnal/bin/console --env=test assets:install Tests/Functionnal/web
   - php Tests/Functionnal/bin/console --env=test assetic:dump
   - nohup php Tests/Functionnal/bin/console --env=test server:run -r vendor/symfony/symfony/src/Symfony/Bundle/FrameworkBundle/Resources/config/router_prod.php &
