--- Bundle/CoreBundle/DataFixtures/ORM/Media/folder.yml
@@ -1,8 +1,0 @@
-Victoire\Bundle\MediaBundle\Entity\Folder:
-    root:
-        name: '/'
-        createdAt: <dateTime('now')>
-        updatedAt: <dateTime('now')>
-        rel: NULL
-        internalName: NULL
-        deleted: 0
