+++ Tests/Features/Context/JavascriptContext.php
@@ -1,0 +1,25 @@
+<?php
+
+namespace Victoire\Tests\Features\Context;
+
+
+use Knp\FriendlyContexts\Context\RawMinkContext;
+
+class JavascriptContext extends RawMinkContext
+{
+    /**
+     * @When /^I maximize the window/
+     */
+    public function iMaximizeTheBroswer()
+    {
+	$this->getSession()->maximizeWindow();
+    }
+
+    /**
+     * @When /^I minimize the window/
+     */
+    public function iMinimizeTheWindow()
+    {
+	$this->getSession()->resizeWindow(500, 500, 'current');
+    }
+}
