+++ Tests/Functionnal/src/Acme/AppBundle/AcmeAppBundle.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Acme\AppBundle;
+
+use Symfony\Component\HttpKernel\Bundle\Bundle;
+
+class AcmeAppBundle extends Bundle
+{
+}
