+++ Tests/Functionnal/src/Acme/AppBundle/DependencyInjection/AcmeAppExtension.php
@@ -1,0 +1,28 @@
+<?php
+
+namespace Acme\AppBundle\DependencyInjection;
+
+use Symfony\Component\DependencyInjection\ContainerBuilder;
+use Symfony\Component\Config\FileLocator;
+use Symfony\Component\HttpKernel\DependencyInjection\Extension;
+use Symfony\Component\DependencyInjection\Loader;
+
+/**
+ * This is the class that loads and manages your bundle configuration
+ *
+ * To learn more see {@link http://symfony.com/doc/current/cookbook/bundles/extension.html}
+ */
+class AcmeAppExtension extends Extension
+{
+    /**
+     * {@inheritdoc}
+     */
+    public function load(array $configs, ContainerBuilder $container)
+    {
+	$configuration = new Configuration();
+	$config = $this->processConfiguration($configuration, $configs);
+
+	$loader = new Loader\XmlFileLoader($container, new FileLocator(__DIR__.'/../Resources/config'));
+	$loader->load('services.xml');
+    }
+}
